+++ packages/sdk/src/prepareOperation.ts
@@ -34,6 +34,7 @@ export async function hasEnoughBalance(
   // check that the user has enough notes to cover the request
   const notes = await notesDB.getNotesFor(asset);
   const balance = notes.reduce((acc, note) => acc + note.value, 0n);
+  console.log(`Have: ${balance}. Requested: ${requestedAmount}`);
   return balance >= requestedAmount;
 }
 
@@ -46,13 +47,8 @@ export async function prepareOperation(
 ): Promise<PreSignOperation> {
   let { refundAddr, maxNumRefunds, gasPrice } = opRequest;
 
-  const {
-    actions,
-    joinSplitRequests,
-    refundAssets,
-    verificationGasLimit,
-    executionGasLimit,
-  } = opRequest;
+  const { actions, joinSplitRequests, refundAssets, executionGasLimit } =
+    opRequest;
 
   // prepare joinSplits
   const joinSplits = (
@@ -81,13 +77,12 @@ export async function prepareOperation(
     maxNumRefunds,
     gasPrice,
 
-    // these may be undefined
-    verificationGasLimit,
+    // this may be undefined
     executionGasLimit,
   };
 
   // simulate if either of the gas limits are undefined
-  const simulationRequired = !verificationGasLimit || !executionGasLimit;
+  const simulationRequired = !executionGasLimit;
   if (simulationRequired) {
     op = await getGasEstimatedOperation(op, walletContract);
   }
@@ -331,8 +326,6 @@ async function getGasEstimatedOperation(
   op: Partial<PreSignOperation>,
   walletContract: Wallet
 ): Promise<PreSignOperation> {
-  op.verificationGasLimit =
-    op.verificationGasLimit ?? DEFAULT_VERIFICATION_GAS_LIMIT;
   op.executionGasLimit = op.executionGasLimit ?? BLOCK_GAS_LIMIT;
   op.gasPrice = op.gasPrice ?? 0n;
 
@@ -346,7 +339,6 @@ async function getGasEstimatedOperation(
   }
   // Give 20% over-estimate
   op.executionGasLimit = (result.executionGas * 12n) / 10n;
-  op.verificationGasLimit = (result.verificationGas + 12n) / 10n;
 
   // since we're simulating, we can get the number of refunds while we're at it
   op.maxNumRefunds = result.numRefunds;